# docs/api/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_anychat_server_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  internal_gateway_handler.createMeetingRequest:
    properties:
      max_participants:
        type: integer
      password:
        type: string
      title:
        type: string
    required:
    - title
    type: object
  internal_gateway_handler.initiateCallRequest:
    properties:
      call_type:
        description: '0-audio/1-video (default: 0)'
        type: integer
      callee_id:
        type: string
    required:
    - callee_id
    type: object
  internal_gateway_handler.joinMeetingRequest:
    properties:
      password:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@anychat.example.com
    name: AnyChat API Support
    url: https://github.com/yzhgit/anychat-server
  description: |-
    AnyChat instant messaging system gateway API service, providing HTTP interfaces for user authentication, user management, and other functions.
    All endpoints requiring authentication must include Authorization: Bearer <token> in the Header.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: AnyChat Gateway API
  version: "1.0"
paths:
...
  /calling/calls:
    get:
      consumes:
      - application/json
      description: Get current user's call history
      parameters:
      - description: page number (default 1)
        in: query
        name: page
        type: integer
      - description: page size (default 20)
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: call logs
      tags:
      - Calling
    post:
      consumes:
      - application/json
      description: Initiate audio/video call to specified user, returns Calling Room
        name and JWT Token
      parameters:
      - description: call request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_gateway_handler.initiateCallRequest'
      produces:
      - application/json
      responses:
        "200":
          description: call initiated successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: initiate call
      tags:
      - Calling
  /calling/calls/{callId}:
    get:
      consumes:
      - application/json
      description: Get detailed info of specified call session
      parameters:
      - description: call ID
        in: path
        name: callId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "404":
          description: call not found
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: get call details
      tags:
      - Calling
  /calling/calls/{callId}/end:
    post:
      consumes:
      - application/json
      description: Caller or callee ends the call
      parameters:
      - description: call ID
        in: path
        name: callId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: end success
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "404":
          description: call not found
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: end call
      tags:
      - Calling
  /calling/calls/{callId}/join:
    post:
      consumes:
      - application/json
      description: Callee accepts call invitation, returns Calling Room name and JWT
        Token
      parameters:
      - description: call ID
        in: path
        name: callId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: answer success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "404":
          description: call not found
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: answer call
      tags:
      - Calling
  /calling/calls/{callId}/reject:
    post:
      consumes:
      - application/json
      description: Callee rejects call invitation
      parameters:
      - description: call ID
        in: path
        name: callId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: reject success
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "404":
          description: call not found
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: reject call
      tags:
      - Calling
  /calling/meetings:
    get:
      consumes:
      - application/json
      description: Get list of currently active meeting rooms
      parameters:
      - description: page number (default 1)
        in: query
        name: page
        type: integer
      - description: page size (default 20)
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: meeting room list
      tags:
      - Calling
    post:
      consumes:
      - application/json
      description: Create new audio/video meeting room, returns meeting info and Calling
        Token
      parameters:
      - description: meeting room info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_gateway_handler.createMeetingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: create success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: create meeting room
      tags:
      - Calling
  /calling/meetings/{roomId}:
    get:
      consumes:
      - application/json
      description: Get detailed info of specified meeting room
      parameters:
      - description: meeting room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "404":
          description: meeting room not found
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: get meeting room
      tags:
      - Calling
  /calling/meetings/{roomId}/end:
    post:
      consumes:
      - application/json
      description: Creator ends meeting room, all participants will be removed after
        room closes
      parameters:
      - description: meeting room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "403":
          description: no permission
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "404":
          description: meeting room not found
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: end meeting room
      tags:
      - Calling
  /calling/meetings/{roomId}/join:
    post:
      consumes:
      - application/json
      description: Join specified meeting room, returns Calling Token
      parameters:
      - description: meeting room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: meeting room password (if any)
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_gateway_handler.joinMeetingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: join success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: parameter error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "403":
          description: wrong password
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "404":
          description: meeting room not found
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
        "500":
          description: server error
          schema:
            $ref: '#/definitions/github_com_anychat_server_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: join meeting room
      tags:
      - Calling
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"